refactor: 출력을 위한 빈 칸이 포함된 체스판 패키지 이동
