+++ src/main/java/chess/Application.java
@@ -2,12 +2,16 @@ package chess;
 
 import chess.controller.ChessController;
 import chess.view.input.InputView;
+import chess.view.output.ConsoleViewChessBoardPositions;
 import chess.view.output.OutputView;
 
 public class Application {
 
     public static void main(String[] args) {
-        ChessController controller = new ChessController(new InputView(), new OutputView());
+        InputView inputView = new InputView();
+        OutputView outputView = new OutputView(ConsoleViewChessBoardPositions.getInstance());
+
+        ChessController controller = new ChessController(inputView, outputView);
         controller.run();
     }
 }
